test(#1491): pair the golden by the SSTable name's BYTES, not a lossy string (W2)
golden_path built the paired golden as format!("{}.jsonl", data_db.display()),
and display() substitutes U+FFFD for every byte that is not valid UTF-8 — so a
*-Data.db whose name is not valid UTF-8 could not be paired with a golden that
IS on disk beside it. That defeated this module's own byte-wise name matching
(fs_probe::name_starts_with/name_ends_with), added a round earlier for exactly
this case: the scan counted such a file correctly and only the pairing failed.

Built by appending to the path's own OsString instead — lossless, and portable
(no std::os::unix).

The lane's other non-diagnostic lossy site went the same way: export() passed
--schema/--data-dir/the output path to the CLI through to_string_lossy(), so a
path that is not valid UTF-8 would be handed to the subprocess as a DIFFERENT
path than the one the test reads back. Now .arg()-chained as OsStr. Every
remaining display()/to_string_lossy() in the lane is inside a diagnostic
message, which is the one place a lossy conversion is correct.

RED-demonstrated: restoring the display() pairing fails the new case with
"no golden <name>.jsonl beside the SSTable it must describe (the directory
holds <name>.jsonl)".
